Copy FBX to Assets Folder

+++ 3DEngine/Source/ModuleSceneIntro.cpp
@@ -390,7 +390,7 @@ void ModuleSceneIntro::LoadScene(const char * file_path)
 	if (loaded_file)
 	{
 		loaded_file.seekg(0, loaded_file.end);
-		buffer_size = loaded_file.tellg();
+		buffer_size = (int)loaded_file.tellg();
 		loaded_file.seekg(0, loaded_file.beg);
 
 		buffer_data = new char[buffer_size];
